entitlement stuff and organizing
