--- app.entitlements
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-  <key>com.apple.security.iokit-user-client-class</key>
-  <array>
-    <string>IOUserClient</string>
-  </array>
-  <key>platform-application</key>
-  <true/>
-  <key>get-task-allow</key>
-  <true/>
-</dict>
-</plist>
