+++ utils/downloader.py
@@ -1,0 +1,54 @@
+import requests
+
+class DpkgDeb:
+    def download_linux_64():
+        res = requests.get("http://ftp.us.debian.org/debian/pool/main/d/dpkg/dpkg_1.20.9_amd64.deb", stream=True)
+        try:
+            if res.status_code == 200:
+                with open(f"dpkg.deb", "wb") as f:
+                    f.write(res.content)
+            else:
+                print(f"[-] dpkg download URL is not reachable. Status code: {res.status_code}")
+                exit(1)
+        except requests.exceptions.RequestException as err:
+            print(f"[-] dpkg download URL is not reachable. Error: {err}")
+            exit(1)
+            
+        subprocess.run(f"ar x dpkg.deb".split(), stdout=subprocess.DEVNULL)
+        subprocess.run(f"tar -xf data.tar.xz".split(), stdout=subprocess.DEVNULL)
+        copy("usr/bin/dpkg-deb", "dpkg-deb")
+        subprocess.run(f"chmod +x dpkg-deb".split(), stdout=subprocess.DEVNULL)
+        os.remove("data.tar.xz")
+        os.remove("control.tar.xz")
+        os.remove("debian-binary")
+        os.remove("dpkg.deb")
+        rmtree("etc")
+        rmtree("sbin")
+        rmtree("usr")
+        rmtree("var")
+        
+    def download_linux_arm64():
+        res = requests.get("http://ftp.us.debian.org/debian/pool/main/d/dpkg/dpkg_1.20.9_arm64.deb", stream=True)
+        try:
+            if res.status_code == 200:
+                with open(f"dpkg.deb", "wb") as f:
+                    f.write(res.content)
+            else:
+                print(f"[-] dpkg download URL is not reachable. Status code: {res.status_code}")
+                exit(1)
+        except requests.exceptions.RequestException as err:
+            print(f"[-] dpkg download URL is not reachable. Error: {err}")
+            exit(1)
+            
+        subprocess.run(f"ar x dpkg.deb".split(), stdout=subprocess.DEVNULL)
+        subprocess.run(f"tar -xf data.tar.xz".split(), stdout=subprocess.DEVNULL)
+        copy("usr/bin/dpkg-deb", "dpkg-deb")
+        subprocess.run(f"chmod +x dpkg-deb".split(), stdout=subprocess.DEVNULL)
+        os.remove("data.tar.xz")
+        os.remove("control.tar.xz")
+        os.remove("debian-binary")
+        os.remove("dpkg.deb")
+        rmtree("etc")
+        rmtree("sbin")
+        rmtree("usr")
+        rmtree("var")
